requirements.txt
torch
torchvision
timm
albumentations
opencv-python
pandas
numpy
scikit-learn
pyyaml
tqdm
